Kind of working
The software is able to analyze some types of tif files. Usually they have to be large files with many frames. I think it has to be at least 20.

tsdata_to_infocrit had issues where the input parameter needed to be positive definite. I created a function that normalized it (chol_reg). Hopefully this does not break the analysis.

The fast_oopsi function had parameter 'b' instead of 'P.b' which was causing some errors